Return value of sccp_rtp_get_sampleRate, is larger than uint8_t, changed to returning int
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@5363 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/sccp_rtp.c
@@ -278,7 +278,7 @@ int sccp_rtp_get_sampleRate(skinny_codec_t codec)
 	if (PBX(rtp_get_sampleRate)) {
 		return PBX(rtp_get_sampleRate) (codec);
 	} else {
-		return (uint8_t) 3840;
+		return 3840;
 	}
 }
